pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"] 

[project]
name = "llmos-cli"
version = "0.1.0"
description = "A tool to deconstruct code repositories into a semantic IR for LLMs."
author = "Fred Bliss"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.1",
    "toml>=0.10.2",
    "tree-sitter>=0.21.3",
    "tree-sitter-languages==1.10.2",
]

[project.urls]
Homepage = "https://github.com/fblissjr/llmos-cli"
Issues = "https://github.com/fblissjr/llmos-cli"

[project.scripts]
llmos-cli = "src.cli:main"